Add "successors" to basic blocks
This allows basicblock to refer to other successor basic blocks with a
"tag" (name).

As each tag may only used once per basic block this uses an
unordered_map to store the successors.

To avoid the circular reference a weak pointer is used rather than the
shared pointer directly.

fmt